Move magic-io-number to correctness
