+++ fortitude/resources/test/fixtures/correctness/C032.f90
